refactor(tests): split apart long test file
