+++ tests/cluster/commands_test.py
@@ -1,12 +1,10 @@
 import datetime
-import re
 import time
 
 import pytest
 
 from tests.cluster.conftest import skip_if_redis_py_version_lt
 from tests.cluster.conftest import skip_if_server_version_lt
-from yaaredis.exceptions import DataError
 from yaaredis.exceptions import RedisClusterException
 from yaaredis.exceptions import RedisError
 from yaaredis.exceptions import ResponseError
@@ -707,1049 +705,3 @@ async def test_type(r):
     await r.delete('a')
     await r.zadd('a', **{'1': 1})
     assert await r.type('a') == b('zset')
-
-
-# LIST COMMANDS
-@pytest.mark.asyncio
-async def test_blpop(r):
-    await r.flushdb()
-    await r.rpush('a{foo}', '1', '2')
-    await r.rpush('b{foo}', '3', '4')
-    assert await r.blpop(['b{foo}', 'a{foo}'], timeout=1) == (b('b{foo}'), b('3'))
-    assert await r.blpop(['b{foo}', 'a{foo}'], timeout=1) == (b('b{foo}'), b('4'))
-    assert await r.blpop(['b{foo}', 'a{foo}'], timeout=1) == (b('a{foo}'), b('1'))
-    assert await r.blpop(['b{foo}', 'a{foo}'], timeout=1) == (b('a{foo}'), b('2'))
-    assert await r.blpop(['b{foo}', 'a{foo}'], timeout=1) is None
-    await r.rpush('c{foo}', '1')
-    assert await r.blpop('c{foo}', timeout=1) == (b('c{foo}'), b('1'))
-
-
-@pytest.mark.asyncio
-async def test_brpop(r):
-    await r.flushdb()
-    await r.rpush('a{foo}', '1', '2')
-    await r.rpush('b{foo}', '3', '4')
-    assert await r.brpop(['b{foo}', 'a{foo}'], timeout=1) == (b('b{foo}'), b('4'))
-    assert await r.brpop(['b{foo}', 'a{foo}'], timeout=1) == (b('b{foo}'), b('3'))
-    assert await r.brpop(['b{foo}', 'a{foo}'], timeout=1) == (b('a{foo}'), b('2'))
-    assert await r.brpop(['b{foo}', 'a{foo}'], timeout=1) == (b('a{foo}'), b('1'))
-    assert await r.brpop(['b{foo}', 'a{foo}'], timeout=1) is None
-    await r.rpush('c{foo}', '1')
-    assert await r.brpop('c{foo}', timeout=1) == (b('c{foo}'), b('1'))
-
-
-@pytest.mark.asyncio
-async def test_brpoplpush(r):
-    await r.flushdb()
-    await r.rpush('a{foo}', '1', '2')
-    await r.rpush('b{foo}', '3', '4')
-    assert await r.brpoplpush('a{foo}', 'b{foo}') == b('2')
-    assert await r.brpoplpush('a{foo}', 'b{foo}') == b('1')
-    assert await r.brpoplpush('a{foo}', 'b{foo}', timeout=1) is None
-    assert await r.lrange('a{foo}', 0, -1) == []
-    assert await r.lrange('b{foo}', 0, -1) == [b('1'), b('2'), b('3'), b('4')]
-
-
-@pytest.mark.asyncio
-async def test_brpoplpush_empty_string(r):
-    await r.flushdb()
-    await r.rpush('a', '')
-    assert await r.brpoplpush('a', 'b') == b('')
-
-
-@pytest.mark.asyncio
-async def test_lindex(r):
-    await r.flushdb()
-    await r.rpush('a', '1', '2', '3')
-    assert await r.lindex('a', '0') == b('1')
-    assert await r.lindex('a', '1') == b('2')
-    assert await r.lindex('a', '2') == b('3')
-
-
-@pytest.mark.asyncio
-async def test_linsert(r):
-    await r.flushdb()
-    await r.rpush('a', '1', '2', '3')
-    assert await r.linsert('a', 'after', '2', '2.5') == 4
-    assert await r.lrange('a', 0, -1) == [b('1'), b('2'), b('2.5'), b('3')]
-    assert await r.linsert('a', 'before', '2', '1.5') == 5
-    assert (await r.lrange('a', 0, -1) == [b('1'), b('1.5'), b('2'), b('2.5'), b('3')])
-
-
-@pytest.mark.asyncio
-async def test_llen(r):
-    await r.flushdb()
-    await r.rpush('a', '1', '2', '3')
-    assert await r.llen('a') == 3
-
-
-@pytest.mark.asyncio
-async def test_lpop(r):
-    await r.flushdb()
-    await r.rpush('a', '1', '2', '3')
-    assert await r.lpop('a') == b('1')
-    assert await r.lpop('a') == b('2')
-    assert await r.lpop('a') == b('3')
-    assert await r.lpop('a') is None
-
-
-@pytest.mark.asyncio
-async def test_lpush(r):
-    await r.flushdb()
-    assert await r.lpush('a', '1') == 1
-    assert await r.lpush('a', '2') == 2
-    assert await r.lpush('a', '3', '4') == 4
-    assert await r.lrange('a', 0, -1) == [b('4'), b('3'), b('2'), b('1')]
-
-
-@pytest.mark.asyncio
-async def test_lpushx(r):
-    await r.flushdb()
-    assert await r.lpushx('a', '1') == 0
-    assert await r.lrange('a', 0, -1) == []
-    await r.rpush('a', '1', '2', '3')
-    assert await r.lpushx('a', '4') == 4
-    assert await r.lrange('a', 0, -1) == [b('4'), b('1'), b('2'), b('3')]
-
-
-@pytest.mark.asyncio
-async def test_lrange(r):
-    await r.flushdb()
-    await r.rpush('a', '1', '2', '3', '4', '5')
-    assert await r.lrange('a', 0, 2) == [b('1'), b('2'), b('3')]
-    assert await r.lrange('a', 2, 10) == [b('3'), b('4'), b('5')]
-    assert await r.lrange('a', 0, -1) == [b('1'), b('2'), b('3'), b('4'), b('5')]
-
-
-@pytest.mark.asyncio
-async def test_lrem(r):
-    await r.flushdb()
-    await r.rpush('a', '1', '1', '1', '1')
-    assert await r.lrem('a', '1', 1) == 1
-    assert await r.lrange('a', 0, -1) == [b('1'), b('1'), b('1')]
-    assert await r.lrem('a', 0, '1') == 3
-    assert await r.lrange('a', 0, -1) == []
-
-
-@pytest.mark.asyncio
-async def test_lset(r):
-    await r.flushdb()
-    await r.rpush('a', '1', '2', '3')
-    assert await r.lrange('a', 0, -1) == [b('1'), b('2'), b('3')]
-    assert await r.lset('a', 1, '4')
-    assert await r.lrange('a', 0, 2) == [b('1'), b('4'), b('3')]
-
-
-@pytest.mark.asyncio
-async def test_ltrim(r):
-    await r.flushdb()
-    await r.rpush('a', '1', '2', '3')
-    assert await r.ltrim('a', 0, 1)
-    assert await r.lrange('a', 0, -1) == [b('1'), b('2')]
-
-
-@pytest.mark.asyncio
-async def test_rpop(r):
-    await r.flushdb()
-    await r.rpush('a', '1', '2', '3')
-    assert await r.rpop('a') == b('3')
-    assert await r.rpop('a') == b('2')
-    assert await r.rpop('a') == b('1')
-    assert await r.rpop('a') is None
-
-
-@pytest.mark.asyncio
-async def test_rpoplpush(r):
-    await r.flushdb()
-    await r.rpush('a', 'a1', 'a2', 'a3')
-    await r.rpush('b', 'b1', 'b2', 'b3')
-    assert await r.rpoplpush('a', 'b') == b('a3')
-    assert await r.lrange('a', 0, -1) == [b('a1'), b('a2')]
-    assert await r.lrange('b', 0, -1) == [b('a3'), b('b1'), b('b2'), b('b3')]
-
-
-@pytest.mark.asyncio
-async def test_rpush(r):
-    await r.flushdb()
-    assert await r.rpush('a', '1') == 1
-    assert await r.rpush('a', '2') == 2
-    assert await r.rpush('a', '3', '4') == 4
-    assert await r.lrange('a', 0, -1) == [b('1'), b('2'), b('3'), b('4')]
-
-
-@pytest.mark.asyncio
-async def test_rpushx(r):
-    await r.flushdb()
-    assert await r.rpushx('a', 'b') == 0
-    assert await r.lrange('a', 0, -1) == []
-    await r.rpush('a', '1', '2', '3')
-    assert await r.rpushx('a', '4') == 4
-    assert await r.lrange('a', 0, -1) == [b('1'), b('2'), b('3'), b('4')]
-
-
-# SCAN COMMANDS
-@pytest.mark.asyncio
-async def test_scan(r):
-    await r.flushdb()
-    await r.set('a', 1)
-    await r.set('b', 2)
-    await r.set('c', 3)
-    keys = []
-    for result in (await r.scan()).values():
-        cursor, partial_keys = result
-        assert cursor == 0
-        keys += partial_keys
-
-    assert set(keys) == {b('a'), b('b'), b('c')}
-
-    keys = []
-    for result in (await r.scan(match='a')).values():
-        cursor, partial_keys = result
-        assert cursor == 0
-        keys += partial_keys
-    assert set(keys) == {b('a')}
-
-
-@pytest.mark.asyncio
-async def test_sscan(r):
-    await r.flushdb()
-    await r.sadd('a', 1, 2, 3)
-    cursor, members = await r.sscan('a')
-    assert cursor == 0
-    assert set(members) == {b('1'), b('2'), b('3')}
-    _, members = await r.sscan('a', match=b('1'))
-    assert set(members) == {b('1')}
-
-
-@pytest.mark.asyncio
-async def test_hscan(r):
-    await r.flushdb()
-    await r.hmset('a', {'a': 1, 'b': 2, 'c': 3})
-    cursor, dic = await r.hscan('a')
-    assert cursor == 0
-    assert dic == {b('a'): b('1'), b('b'): b('2'), b('c'): b('3')}
-    _, dic = await r.hscan('a', match='a')
-    assert dic == {b('a'): b('1')}
-
-
-@pytest.mark.asyncio
-async def test_zscan(r):
-    await r.flushdb()
-    await r.zadd('a', 1, 'a', 2, 'b', 3, 'c')
-    cursor, pairs = await r.zscan('a')
-    assert cursor == 0
-    assert set(pairs) == {(b('a'), 1), (b('b'), 2), (b('c'), 3)}
-    _, pairs = await r.zscan('a', match='a')
-    assert set(pairs) == {(b('a'), 1)}
-
-
-# SET COMMANDS
-@pytest.mark.asyncio
-async def test_sadd(r):
-    await r.flushdb()
-    members = {b('1'), b('2'), b('3')}
-    await r.sadd('a', *members)
-    assert await r.smembers('a') == members
-
-
-@pytest.mark.asyncio
-async def test_scard(r):
-    await r.flushdb()
-    await r.sadd('a', '1', '2', '3')
-    assert await r.scard('a') == 3
-
-
-@pytest.mark.asyncio
-async def test_sdiff(r):
-    await r.flushdb()
-    await r.sadd('a{foo}', '1', '2', '3')
-    assert await r.sdiff('a{foo}', 'b{foo}') == {b('1'), b('2'), b('3')}
-    await r.sadd('b{foo}', '2', '3')
-    assert await r.sdiff('a{foo}', 'b{foo}') == {b('1')}
-
-
-@pytest.mark.asyncio
-async def test_sdiffstore(r):
-    await r.flushdb()
-    await r.sadd('a{foo}', '1', '2', '3')
-    assert await r.sdiffstore('c{foo}', 'a{foo}', 'b{foo}') == 3
-    assert await r.smembers('c{foo}') == {b('1'), b('2'), b('3')}
-    await r.sadd('b{foo}', '2', '3')
-    assert await r.sdiffstore('c{foo}', 'a{foo}', 'b{foo}') == 1
-    assert await r.smembers('c{foo}') == {b('1')}
-
-    # Diff:s that return empty set should not fail
-    assert await r.sdiffstore('d{foo}', 'e{foo}') == 0
-
-
-@pytest.mark.asyncio
-async def test_sinter(r):
-    await r.flushdb()
-    await r.sadd('a{foo}', '1', '2', '3')
-    assert await r.sinter('a{foo}', 'b{foo}') == set()
-    await r.sadd('b{foo}', '2', '3')
-    assert await r.sinter('a{foo}', 'b{foo}') == {b('2'), b('3')}
-
-
-@pytest.mark.asyncio
-async def test_sinterstore(r):
-    await r.flushdb()
-    await r.sadd('a{foo}', '1', '2', '3')
-    assert await r.sinterstore('c{foo}', 'a{foo}', 'b{foo}') == 0
-    assert await r.smembers('c{foo}') == set()
-    await r.sadd('b{foo}', '2', '3')
-    assert await r.sinterstore('c{foo}', 'a{foo}', 'b{foo}') == 2
-    assert await r.smembers('c{foo}') == {b('2'), b('3')}
-
-
-@pytest.mark.asyncio
-async def test_sismember(r):
-    await r.flushdb()
-    await r.sadd('a', '1', '2', '3')
-    assert await r.sismember('a', '1')
-    assert await r.sismember('a', '2')
-    assert await r.sismember('a', '3')
-    assert not await r.sismember('a', '4')
-
-
-@pytest.mark.asyncio
-async def test_smembers(r):
-    await r.flushdb()
-    await r.sadd('a', '1', '2', '3')
-    assert await r.smembers('a') == {b('1'), b('2'), b('3')}
-
-
-@pytest.mark.asyncio
-async def test_smove(r):
-    await r.flushdb()
-    await r.sadd('a{foo}', 'a1', 'a2')
-    await r.sadd('b{foo}', 'b1', 'b2')
-    assert await r.smove('a{foo}', 'b{foo}', 'a1')
-    assert await r.smembers('a{foo}') == {b('a2')}
-    assert await r.smembers('b{foo}') == {b('b1'), b('b2'), b('a1')}
-
-
-@pytest.mark.asyncio
-async def test_spop(r):
-    await r.flushdb()
-    s = [b('1'), b('2'), b('3')]
-    await r.sadd('a', *s)
-    value = await r.spop('a')
-    assert value in s
-    assert await r.smembers('a') == set(s) - {value}
-
-
-@pytest.mark.asyncio
-async def test_srandmember(r):
-    await r.flushdb()
-    s = [b('1'), b('2'), b('3')]
-    await r.sadd('a', *s)
-    assert await r.srandmember('a') in s
-
-
-@pytest.mark.asyncio
-async def test_srandmember_multi_value(r):
-    await r.flushdb()
-    s = [b('1'), b('2'), b('3')]
-    await r.sadd('a', *s)
-    randoms = await r.srandmember('a', number=2)
-    assert len(randoms) == 2
-    assert set(randoms).intersection(s) == set(randoms)
-
-
-@pytest.mark.asyncio
-async def test_srem(r):
-    await r.flushdb()
-    await r.sadd('a', '1', '2', '3', '4')
-    assert await r.srem('a', '5') == 0
-    assert await r.srem('a', '2', '4') == 2
-    assert await r.smembers('a') == {b('1'), b('3')}
-
-
-@pytest.mark.asyncio
-async def test_sunion(r):
-    await r.flushdb()
-    await r.sadd('a{foo}', '1', '2')
-    await r.sadd('b{foo}', '2', '3')
-    assert await r.sunion('a{foo}', 'b{foo}') == {b('1'), b('2'), b('3')}
-
-
-@pytest.mark.asyncio
-async def test_sunionstore(r):
-    await r.flushdb()
-    await r.sadd('a{foo}', '1', '2')
-    await r.sadd('b{foo}', '2', '3')
-    assert await r.sunionstore('c{foo}', 'a{foo}', 'b{foo}') == 3
-    assert await r.smembers('c{foo}') == {b('1'), b('2'), b('3')}
-
-
-# SORTED SET COMMANDS
-@pytest.mark.asyncio
-async def test_zadd(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3)
-    assert await r.zrange('a', 0, -1) == [b('a1'), b('a2'), b('a3')]
-
-
-@pytest.mark.asyncio
-async def test_zaddoption(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1)
-    assert int(await r.zscore('a', 'a1')) == 1
-    assert int(await r.zaddoption('a', 'NX', a1=2)) == 0
-    assert int(await r.zaddoption('a', 'NX CH', a1=2)) == 0
-    assert int(await r.zscore('a', 'a1')) == 1
-    assert await r.zcard('a') == 1
-    assert int(await r.zaddoption('a', 'XX', a2=1)) == 0
-    assert await r.zcard('a') == 1
-    assert int(await r.zaddoption('a', 'XX', a1=2)) == 0
-    assert int(await r.zaddoption('a', 'XX CH', a1=3)) == 1
-    assert int(await r.zscore('a', 'a1')) == 3
-    assert int(await r.zaddoption('a', 'NX', a2=1)) == 1
-    assert int(await r.zaddoption('a', 'NX CH', a3=1)) == 1
-    assert await r.zcard('a') == 3
-    await r.zaddoption('a', 'INCR', a3=1)
-    assert int(await r.zscore('a', 'a3')) == 2
-
-
-@pytest.mark.asyncio
-async def test_zcard(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3)
-    assert await r.zcard('a') == 3
-
-
-@pytest.mark.asyncio
-async def test_zcount(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3)
-    assert await r.zcount('a', '-inf', '+inf') == 3
-    assert await r.zcount('a', 1, 2) == 2
-    assert await r.zcount('a', 10, 20) == 0
-
-
-@pytest.mark.asyncio
-async def test_zincrby(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3)
-    assert await r.zincrby('a', 'a2') == 3.0
-    assert await r.zincrby('a', 'a3', amount=5) == 8.0
-    assert await r.zscore('a', 'a2') == 3.0
-    assert await r.zscore('a', 'a3') == 8.0
-
-
-@pytest.mark.asyncio
-async def test_zlexcount(r):
-    await r.flushdb()
-    await r.zadd('a', a=0, b=0, c=0, d=0, e=0, f=0, g=0)
-    assert await r.zlexcount('a', '-', '+') == 7
-    assert await r.zlexcount('a', '[b', '[f') == 5
-
-
-@pytest.mark.asyncio
-async def test_zinterstore_fail_cross_slot(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=1, a3=1)
-    await r.zadd('b', a1=2, a2=2, a3=2)
-    await r.zadd('c', a1=6, a3=5, a4=4)
-    with pytest.raises(ResponseError) as excinfo:
-        await r.zinterstore('d', ['a', 'b', 'c'])
-    assert re.search('ClusterCrossSlotError', str(excinfo))
-
-
-@pytest.mark.asyncio
-async def test_zinterstore_sum(r):
-    await r.flushdb()
-    await r.zadd('a{foo}', a1=1, a2=1, a3=1)
-    await r.zadd('b{foo}', a1=2, a2=2, a3=2)
-    await r.zadd('c{foo}', a1=6, a3=5, a4=4)
-    assert await r.zinterstore('d{foo}', ['a{foo}', 'b{foo}', 'c{foo}']) == 2
-    assert (await r.zrange('d{foo}', 0, -1, withscores=True) == [(b('a3'), 8), (b('a1'), 9)])
-
-
-@pytest.mark.asyncio
-async def test_zinterstore_max(r):
-    await r.flushdb()
-    await r.zadd('a{foo}', a1=1, a2=1, a3=1)
-    await r.zadd('b{foo}', a1=2, a2=2, a3=2)
-    await r.zadd('c{foo}', a1=6, a3=5, a4=4)
-    assert await r.zinterstore('d{foo}', ['a{foo}', 'b{foo}', 'c{foo}'], aggregate='MAX') == 2
-    assert (await r.zrange('d{foo}', 0, -1, withscores=True) == [(b('a3'), 5), (b('a1'), 6)])
-
-
-@pytest.mark.asyncio
-async def test_zinterstore_min(r):
-    await r.flushdb()
-    await r.zadd('a{foo}', a1=1, a2=2, a3=3)
-    await r.zadd('b{foo}', a1=2, a2=3, a3=5)
-    await r.zadd('c{foo}', a1=6, a3=5, a4=4)
-    assert await r.zinterstore('d{foo}', ['a{foo}', 'b{foo}', 'c{foo}'], aggregate='MIN') == 2
-    assert (await r.zrange('d{foo}', 0, -1, withscores=True) == [(b('a1'), 1), (b('a3'), 3)])
-
-
-@pytest.mark.asyncio
-async def test_zinterstore_with_weight(r):
-    await r.flushdb()
-    await r.zadd('a{foo}', a1=1, a2=1, a3=1)
-    await r.zadd('b{foo}', a1=2, a2=2, a3=2)
-    await r.zadd('c{foo}', a1=6, a3=5, a4=4)
-    assert await r.zinterstore('d{foo}', {'a{foo}': 1, 'b{foo}': 2, 'c{foo}': 3}) == 2
-    assert (await r.zrange('d{foo}', 0, -1, withscores=True) == [(b('a3'), 20), (b('a1'), 23)])
-
-
-@pytest.mark.asyncio
-async def test_zrange(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3)
-    assert await r.zrange('a', 0, 1) == [b('a1'), b('a2')]
-    assert await r.zrange('a', 1, 2) == [b('a2'), b('a3')]
-
-    # withscores
-    assert (await r.zrange('a', 0, 1, withscores=True) == [(b('a1'), 1.0), (b('a2'), 2.0)])
-    assert (await r.zrange('a', 1, 2, withscores=True) == [(b('a2'), 2.0), (b('a3'), 3.0)])
-
-    # custom score function
-    assert (await r.zrange('a', 0, 1, withscores=True, score_cast_func=int) == [(b('a1'), 1), (b('a2'), 2)])
-
-
-@pytest.mark.asyncio
-async def test_zrangebylex(r):
-    await r.flushdb()
-    await r.zadd('a', a=0, b=0, c=0, d=0, e=0, f=0, g=0)
-    assert await r.zrangebylex('a', '-', '[c') == [b('a'), b('b'), b('c')]
-    assert await r.zrangebylex('a', '-', '(c') == [b('a'), b('b')]
-    assert (await r.zrangebylex('a', '[aaa', '(g') == [b('b'), b('c'), b('d'), b('e'), b('f')])
-    assert await r.zrangebylex('a', '[f', '+') == [b('f'), b('g')]
-    assert await r.zrangebylex('a', '-', '+', start=3, num=2) == [b('d'), b('e')]
-
-
-@pytest.mark.asyncio
-async def test_zrangebyscore(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3, a4=4, a5=5)
-    assert await r.zrangebyscore('a', 2, 4) == [b('a2'), b('a3'), b('a4')]
-
-    # slicing with start/num
-    expected = [b('a3'), b('a4')]
-    assert await r.zrangebyscore('a', 2, 4, start=1, num=2) == expected
-
-    # withscores
-    expected = [(b('a2'), 2.0), (b('a3'), 3.0), (b('a4'), 4.0)]
-    assert await r.zrangebyscore('a', 2, 4, withscores=True) == expected
-
-    # custom score function
-    expected = [(b('a2'), 2), (b('a3'), 3), (b('a4'), 4)]
-    assert await r.zrangebyscore('a', 2, 4, withscores=True,
-                                 score_cast_func=int) == expected
-
-
-@pytest.mark.asyncio
-async def test_zrank(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3, a4=4, a5=5)
-    assert await r.zrank('a', 'a1') == 0
-    assert await r.zrank('a', 'a2') == 1
-    assert await r.zrank('a', 'a6') is None
-
-
-@pytest.mark.asyncio
-async def test_zrem(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3)
-    assert await r.zrem('a', 'a2') == 1
-    assert await r.zrange('a', 0, -1) == [b('a1'), b('a3')]
-    assert await r.zrem('a', 'b') == 0
-    assert await r.zrange('a', 0, -1) == [b('a1'), b('a3')]
-
-
-@pytest.mark.asyncio
-async def test_zrem_multiple_keys(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3)
-    assert await r.zrem('a', 'a1', 'a2') == 2
-    assert await r.zrange('a', 0, 5) == [b('a3')]
-
-
-@pytest.mark.asyncio
-async def test_zremrangebylex(r):
-    await r.flushdb()
-    await r.zadd('a', a=0, b=0, c=0, d=0, e=0, f=0, g=0)
-    assert await r.zremrangebylex('a', '-', '[c') == 3
-    assert await r.zrange('a', 0, -1) == [b('d'), b('e'), b('f'), b('g')]
-    assert await r.zremrangebylex('a', '[f', '+') == 2
-    assert await r.zrange('a', 0, -1) == [b('d'), b('e')]
-    assert await r.zremrangebylex('a', '[h', '+') == 0
-    assert await r.zrange('a', 0, -1) == [b('d'), b('e')]
-
-
-@pytest.mark.asyncio
-async def test_zremrangebyrank(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3, a4=4, a5=5)
-    assert await r.zremrangebyrank('a', 1, 3) == 3
-    assert await r.zrange('a', 0, 5) == [b('a1'), b('a5')]
-
-
-@pytest.mark.asyncio
-async def test_zremrangebyscore(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3, a4=4, a5=5)
-    assert await r.zremrangebyscore('a', 2, 4) == 3
-    assert await r.zrange('a', 0, -1) == [b('a1'), b('a5')]
-    assert await r.zremrangebyscore('a', 2, 4) == 0
-    assert await r.zrange('a', 0, -1) == [b('a1'), b('a5')]
-
-
-@pytest.mark.asyncio
-async def test_zrevrange(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3)
-    assert await r.zrevrange('a', 0, 1) == [b('a3'), b('a2')]
-    assert await r.zrevrange('a', 1, 2) == [b('a2'), b('a1')]
-
-    # withscores
-    expected = [(b('a3'), 3.0), (b('a2'), 2.0)]
-    assert await r.zrevrange('a', 0, 1, withscores=True) == expected
-    expected = [(b('a2'), 2.0), (b('a1'), 1.0)]
-    assert await r.zrevrange('a', 1, 2, withscores=True) == expected
-
-    # custom score function
-    expected = [(b('a3'), 3.0), (b('a2'), 2.0)]
-    assert await r.zrevrange('a', 0, 1, withscores=True,
-                             score_cast_func=int) == expected
-
-
-@pytest.mark.asyncio
-async def test_zrevrangebyscore(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3, a4=4, a5=5)
-    assert await r.zrevrangebyscore('a', 4, 2) == [b('a4'), b('a3'), b('a2')]
-
-    # slicing with start/num
-    expected = [b('a3'), b('a2')]
-    assert await r.zrevrangebyscore('a', 4, 2, start=1, num=2) == expected
-
-    # withscores
-    expected = [(b('a4'), 4.0), (b('a3'), 3.0), (b('a2'), 2.0)]
-    assert await r.zrevrangebyscore('a', 4, 2, withscores=True) == expected
-
-    # custom score function
-    expected = [(b('a4'), 4), (b('a3'), 3), (b('a2'), 2)]
-    assert await r.zrevrangebyscore('a', 4, 2, withscores=True,
-                                    score_cast_func=int) == expected
-
-
-@pytest.mark.asyncio
-async def test_zrevrank(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3, a4=4, a5=5)
-    assert await r.zrevrank('a', 'a1') == 4
-    assert await r.zrevrank('a', 'a2') == 3
-    assert await r.zrevrank('a', 'a6') is None
-
-
-@pytest.mark.asyncio
-async def test_zscore(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=2, a3=3)
-    assert await r.zscore('a', 'a1') == 1.0
-    assert await r.zscore('a', 'a2') == 2.0
-    assert await r.zscore('a', 'a4') is None
-
-
-@pytest.mark.asyncio
-async def test_zunionstore_fail_crossslot(r):
-    await r.flushdb()
-    await r.zadd('a', a1=1, a2=1, a3=1)
-    await r.zadd('b', a1=2, a2=2, a3=2)
-    await r.zadd('c', a1=6, a3=5, a4=4)
-    with pytest.raises(ResponseError) as excinfo:
-        await r.zunionstore('d', ['a', 'b', 'c'])
-    assert re.search('ClusterCrossSlotError', str(excinfo))
-
-
-@pytest.mark.asyncio
-async def test_zunionstore_sum(r):
-    await r.flushdb()
-    await r.zadd('a{foo}', a1=1, a2=1, a3=1)
-    await r.zadd('b{foo}', a1=2, a2=2, a3=2)
-    await r.zadd('c{foo}', a1=6, a3=5, a4=4)
-    assert await r.zunionstore('d{foo}', ['a{foo}', 'b{foo}', 'c{foo}']) == 4
-    expected = [(b('a2'), 3), (b('a4'), 4), (b('a3'), 8), (b('a1'), 9)]
-    assert await r.zrange('d{foo}', 0, -1, withscores=True) == expected
-
-
-@pytest.mark.asyncio
-async def test_zunionstore_max(r):
-    await r.flushdb()
-    await r.zadd('a{foo}', a1=1, a2=1, a3=1)
-    await r.zadd('b{foo}', a1=2, a2=2, a3=2)
-    await r.zadd('c{foo}', a1=6, a3=5, a4=4)
-    assert await r.zunionstore('d{foo}', ['a{foo}', 'b{foo}', 'c{foo}'], aggregate='MAX') == 4
-    expected = [(b('a2'), 2), (b('a4'), 4), (b('a3'), 5), (b('a1'), 6)]
-    assert await r.zrange('d{foo}', 0, -1, withscores=True) == expected
-
-
-@pytest.mark.asyncio
-async def test_zunionstore_min(r):
-    await r.flushdb()
-    await r.zadd('a{foo}', a1=1, a2=2, a3=3)
-    await r.zadd('b{foo}', a1=2, a2=2, a3=4)
-    await r.zadd('c{foo}', a1=6, a3=5, a4=4)
-    assert await r.zunionstore('d{foo}', ['a{foo}', 'b{foo}', 'c{foo}'], aggregate='MIN') == 4
-    expected = [(b('a1'), 1), (b('a2'), 2), (b('a3'), 3), (b('a4'), 4)]
-    assert await r.zrange('d{foo}', 0, -1, withscores=True) == expected
-
-
-@pytest.mark.asyncio
-async def test_zunionstore_with_weight(r):
-    await r.flushdb()
-    await r.zadd('a{foo}', a1=1, a2=1, a3=1)
-    await r.zadd('b{foo}', a1=2, a2=2, a3=2)
-    await r.zadd('c{foo}', a1=6, a3=5, a4=4)
-    assert await r.zunionstore('d{foo}', {'a{foo}': 1, 'b{foo}': 2, 'c{foo}': 3}) == 4
-    expected = [(b('a2'), 5), (b('a4'), 12), (b('a3'), 20), (b('a1'), 23)]
-    assert await r.zrange('d{foo}', 0, -1, withscores=True) == expected
-
-
-# HYPERLOGLOG TESTS
-@pytest.mark.asyncio
-async def test_pfadd(r):
-    await r.flushdb()
-    members = {b('1'), b('2'), b('3')}
-    assert await r.pfadd('a', *members) == 1
-    assert await r.pfadd('a', *members) == 0
-    assert await r.pfcount('a') == len(members)
-
-
-@pytest.mark.asyncio
-@pytest.mark.xfail(reason='New pfcount in 2.10.5 currently breaks in cluster')
-@skip_if_server_version_lt('2.8.9')
-async def test_pfcount(r):
-    await r.flushdb()
-    members = {b('1'), b('2'), b('3')}
-    await r.pfadd('a', *members)
-    assert await r.pfcount('a') == len(members)
-    members_b = {b('2'), b('3'), b('4')}
-    await r.pfadd('b', *members_b)
-    assert await r.pfcount('b') == len(members_b)
-    assert await r.pfcount('a', 'b') == len(members_b.union(members))
-
-
-@pytest.mark.asyncio
-async def test_pfmerge(r):
-    await r.flushdb()
-    mema = {b('1'), b('2'), b('3')}
-    memb = {b('2'), b('3'), b('4')}
-    memc = {b('5'), b('6'), b('7')}
-    await r.pfadd('a', *mema)
-    await r.pfadd('b', *memb)
-    await r.pfadd('c', *memc)
-    await r.pfmerge('d', 'c', 'a')
-    assert await r.pfcount('d') == 6
-    await r.pfmerge('d', 'b')
-    assert await r.pfcount('d') == 7
-
-
-# HASH COMMANDS
-@pytest.mark.asyncio
-async def test_hget_and_hset(r):
-    await r.flushdb()
-    await r.hmset('a', {'1': 1, '2': 2, '3': 3})
-    assert await r.hget('a', '1') == b('1')
-    assert await r.hget('a', '2') == b('2')
-    assert await r.hget('a', '3') == b('3')
-
-    # field was updated, redis returns 0
-    assert await r.hset('a', '2', 5) == 0
-    assert await r.hget('a', '2') == b('5')
-
-    # field is new, redis returns 1
-    assert await r.hset('a', '4', 4) == 1
-    assert await r.hget('a', '4') == b('4')
-
-    # key inside of hash that doesn't exist returns null value
-    assert await r.hget('a', 'b') is None
-
-
-@pytest.mark.asyncio
-async def test_hdel(r):
-    await r.flushdb()
-    await r.hmset('a', {'1': 1, '2': 2, '3': 3})
-    assert await r.hdel('a', '2') == 1
-    assert await r.hget('a', '2') is None
-    assert await r.hdel('a', '1', '3') == 2
-    assert await r.hlen('a') == 0
-
-
-@pytest.mark.asyncio
-async def test_hexists(r):
-    await r.flushdb()
-    await r.hmset('a', {'1': 1, '2': 2, '3': 3})
-    assert await r.hexists('a', '1')
-    assert not await r.hexists('a', '4')
-
-
-@pytest.mark.asyncio
-async def test_hgetall(r):
-    await r.flushdb()
-    h = {b('a1'): b('1'), b('a2'): b('2'), b('a3'): b('3')}
-    await r.hmset('a', h)
-    assert await r.hgetall('a') == h
-
-
-@pytest.mark.asyncio
-async def test_hincrby(r):
-    await r.flushdb()
-    assert await r.hincrby('a', '1') == 1
-    assert await r.hincrby('a', '1', amount=2) == 3
-    assert await r.hincrby('a', '1', amount=-2) == 1
-
-
-@pytest.mark.asyncio
-async def test_hincrbyfloat(r):
-    await r.flushdb()
-    assert await r.hincrbyfloat('a', '1') == 1.0
-    assert await r.hincrbyfloat('a', '1') == 2.0
-    assert await r.hincrbyfloat('a', '1', 1.2) == 3.2
-
-
-@pytest.mark.asyncio
-async def test_hkeys(r):
-    await r.flushdb()
-    h = {b('a1'): b('1'), b('a2'): b('2'), b('a3'): b('3')}
-    await r.hmset('a', h)
-    local_keys = list(iter(h.keys()))
-    remote_keys = await r.hkeys('a')
-    assert sorted(local_keys) == sorted(remote_keys)
-
-
-@pytest.mark.asyncio
-async def test_hlen(r):
-    await r.flushdb()
-    await r.hmset('a', {'1': 1, '2': 2, '3': 3})
-    assert await r.hlen('a') == 3
-
-
-@pytest.mark.asyncio
-async def test_hmget(r):
-    await r.flushdb()
-    assert await r.hmset('a', {'a': 1, 'b': 2, 'c': 3})
-    assert await r.hmget('a', 'a', 'b', 'c') == [b('1'), b('2'), b('3')]
-
-
-@pytest.mark.asyncio
-async def test_hmset(r):
-    await r.flushdb()
-    h = {b('a'): b('1'), b('b'): b('2'), b('c'): b('3')}
-    assert await r.hmset('a', h)
-    assert await r.hgetall('a') == h
-
-
-@pytest.mark.asyncio
-async def test_hsetnx(r):
-    await r.flushdb()
-    # Initially set the hash field
-    assert await r.hsetnx('a', '1', 1)
-    assert await r.hget('a', '1') == b('1')
-    assert not await r.hsetnx('a', '1', 2)
-    assert await r.hget('a', '1') == b('1')
-
-
-@pytest.mark.asyncio
-async def test_hvals(r):
-    await r.flushdb()
-    h = {b('a1'): b('1'), b('a2'): b('2'), b('a3'): b('3')}
-    await r.hmset('a', h)
-    local_vals = list(iter(h.values()))
-    remote_vals = await r.hvals('a')
-    assert sorted(local_vals) == sorted(remote_vals)
-
-
-# SCAN
-@pytest.mark.asyncio
-async def test_scan_iter(r):
-    await r.flushdb()
-    await r.set('a', 1)
-    await r.set('b', 2)
-    await r.set('c', 3)
-    keys = set()
-    async for key in r.scan_iter():
-        keys.add(key)
-    assert keys == {b('a'), b('b'), b('c')}
-    async for key in r.scan_iter(match='a'):
-        assert key == b('a')
-
-
-@pytest.mark.asyncio
-async def test_scan_iter_multi_page(r):
-    await r.flushdb()
-    await r.set('a', 1)
-    await r.set('b', 2)
-    await r.set('c', 3)
-    await r.set('d', 4)
-    await r.set('e', 5)
-    keys = set()
-    async for key in r.scan_iter(count=1):
-        # three nodes, five items, we've gotta be using a cursor
-        keys.add(key)
-    assert keys == {b('a'), b('b'), b('c'), b('d'), b('e')}
-
-
-# SORT
-@pytest.mark.asyncio
-async def test_sort_basic(r):
-    await r.flushdb()
-    await r.rpush('a', '3', '2', '1', '4')
-    assert await r.sort('a') == [b('1'), b('2'), b('3'), b('4')]
-
-
-@pytest.mark.asyncio
-async def test_sort_limited(r):
-    await r.flushdb()
-    await r.rpush('a', '3', '2', '1', '4')
-    assert await r.sort('a', start=1, num=2) == [b('2'), b('3')]
-
-
-@pytest.mark.asyncio
-async def test_sort_by(r):
-    await r.flushdb()
-    await r.set('score:1', 8)
-    await r.set('score:2', 3)
-    await r.set('score:3', 5)
-    await r.rpush('a', '3', '2', '1')
-    assert await r.sort('a', by='score:*') == [b('2'), b('3'), b('1')]
-
-
-@pytest.mark.asyncio
-async def test_sort_get(r):
-    await r.flushdb()
-    await r.set('user:1', 'u1')
-    await r.set('user:2', 'u2')
-    await r.set('user:3', 'u3')
-    await r.rpush('a', '2', '3', '1')
-    assert await r.sort('a', get='user:*') == [b('u1'), b('u2'), b('u3')]
-
-
-@pytest.mark.asyncio
-async def test_sort_get_multi(r):
-    await r.flushdb()
-    await r.set('user:1', 'u1')
-    await r.set('user:2', 'u2')
-    await r.set('user:3', 'u3')
-    await r.rpush('a', '2', '3', '1')
-    expected = [b('u1'), b('1'), b('u2'), b('2'), b('u3'), b('3')]
-    assert await r.sort('a', get=('user:*', '#')) == expected
-
-
-@pytest.mark.asyncio
-async def test_sort_get_groups_two(r):
-    await r.flushdb()
-    await r.set('user:1', 'u1')
-    await r.set('user:2', 'u2')
-    await r.set('user:3', 'u3')
-    await r.rpush('a', '2', '3', '1')
-    expected = [(b('u1'), b('1')), (b('u2'), b('2')), (b('u3'), b('3'))]
-    assert await r.sort('a', get=('user:*', '#'), groups=True) == expected
-
-
-@pytest.mark.asyncio
-async def test_sort_groups_string_get(r):
-    await r.flushdb()
-    await r.set('user:1', 'u1')
-    await r.set('user:2', 'u2')
-    await r.set('user:3', 'u3')
-    await r.rpush('a', '2', '3', '1')
-    with pytest.raises(DataError):
-        await r.sort('a', get='user:*', groups=True)
-
-
-@pytest.mark.asyncio
-async def test_sort_groups_just_one_get(r):
-    await r.flushdb()
-    await r.set('user:1', 'u1')
-    await r.set('user:2', 'u2')
-    await r.set('user:3', 'u3')
-    await r.rpush('a', '2', '3', '1')
-    with pytest.raises(DataError):
-        await r.sort('a', get=['user:*'], groups=True)
-
-
-@pytest.mark.asyncio
-async def test_sort_groups_no_get(r):
-    await r.flushdb()
-    await r.set('user:1', 'u1')
-    await r.set('user:2', 'u2')
-    await r.set('user:3', 'u3')
-    await r.rpush('a', '2', '3', '1')
-    with pytest.raises(DataError):
-        await r.sort('a', groups=True)
-
-
-@pytest.mark.asyncio
-async def test_sort_groups_three_gets(r):
-    await r.flushdb()
-    await r.set('user:1', 'u1')
-    await r.set('user:2', 'u2')
-    await r.set('user:3', 'u3')
-    await r.set('door:1', 'd1')
-    await r.set('door:2', 'd2')
-    await r.set('door:3', 'd3')
-    await r.rpush('a', '2', '3', '1')
-    assert await r.sort('a', get=('user:*', 'door:*', '#'), groups=True) == [
-        (b('u1'), b('d1'), b('1')),
-        (b('u2'), b('d2'), b('2')),
-        (b('u3'), b('d3'), b('3')),
-    ]
-
-
-@pytest.mark.asyncio
-async def test_sort_desc(r):
-    await r.flushdb()
-    await r.rpush('a', '2', '3', '1')
-    assert await r.sort('a', desc=True) == [b('3'), b('2'), b('1')]
-
-
-@pytest.mark.asyncio
-async def test_sort_alpha(r):
-    await r.flushdb()
-    await r.rpush('a', 'e', 'c', 'b', 'd', 'a')
-    assert (await r.sort('a', alpha=True) == [b('a'), b('b'), b('c'), b('d'), b('e')])
-
-
-@pytest.mark.asyncio
-async def test_sort_store(r):
-    await r.flushdb()
-    await r.rpush('a', '2', '3', '1')
-    assert await r.sort('a', store='sorted_values') == 3
-    assert await r.lrange('sorted_values', 0, -1) == [b('1'), b('2'), b('3')]
-
-
-@pytest.mark.asyncio
-async def test_sort_all_options(r):
-    await r.flushdb()
-    await r.set('user:1:username', 'zeus')
-    await r.set('user:2:username', 'titan')
-    await r.set('user:3:username', 'hermes')
-    await r.set('user:4:username', 'hercules')
-    await r.set('user:5:username', 'apollo')
-    await r.set('user:6:username', 'athena')
-    await r.set('user:7:username', 'hades')
-    await r.set('user:8:username', 'dionysus')
-
-    await r.set('user:1:favorite_drink', 'yuengling')
-    await r.set('user:2:favorite_drink', 'rum')
-    await r.set('user:3:favorite_drink', 'vodka')
-    await r.set('user:4:favorite_drink', 'milk')
-    await r.set('user:5:favorite_drink', 'pinot noir')
-    await r.set('user:6:favorite_drink', 'water')
-    await r.set('user:7:favorite_drink', 'gin')
-    await r.set('user:8:favorite_drink', 'apple juice')
-
-    await r.rpush('gods', '5', '8', '3', '1', '2', '7', '6', '4')
-    num = await r.sort('gods', start=2, num=4, by='user:*:username',
-                       get='user:*:favorite_drink', desc=True, alpha=True,
-                       store='sorted')
-    assert num == 4
-    expected = [b('vodka'), b('milk'), b('gin'), b('apple juice')]
-    assert await r.lrange('sorted', 0, 10) == expected
